Minimal pixi build, fix window size, refactor file names.
