+++ src/main/lib/nvimCommand.js
