+++ src/main/lib/nvimVersion.js
