+++ src/main/lib/shellEnv.js
