+++ src/renderer/input/insertSymbols.js
